Introduce fragment/ schemas for common and http
This allows us to distinguish between fragment schemas that
are meant to be included in final event schemas, and those
that are real event schemas.

We will move all 'mediawiki/**common' schemas into fragment/.

This commit also adds the http field to the mediawiki/client/error schema, and
removes user_agent field, as this will be set in the http.request_headers map.
(We can do this still at 1.0.0 because mediawiki/client/error is not yet used.)

Bug: T242363
Change-Id: I9432f3b1598c2327e14d92e8a6d4b76002aab47f