Add mediawiki/revision/recommendation-create
Adds a new event schema, mediawiki/revision/recommendation-create,
to identify that the given MediaWiki revision now has a recommendation
available. (The recommendation itself is not recorded as part of this
event; the primary purpose is to notify the search index which tracks
which pages have recommendations of a given type.)

The 'performer' fields are somewhat meaningless here (they will hold
the user who created the revision; creating the recommendation is
not tied to a user) but they are a standard part of all MediaWiki
events so it is probably less hassle to include them.

Bug: T261407
Change-Id: I5c18f0d7a5ed858a29121f301f6c82080c37ed17