Rematerialize all schemas wiith enforced numeric bounds
All current versions have been rematerialized with
maximum and minimum values for all integer and number fields.

Non current schema versions had to be manually edited to add
maximum and minimum values to get them to pass tests.

This change looks huge because jsonschema-tools now enforces
a consistent ordering of fields in materialized schema files,
but the only real modification is the addition of maximum and
minimum where it didn't already exist on numeric fields.

Bug: T258659
Change-Id: I52ff95bad81bf3757e03403550855967cc109ed0