development/mediawiki/page/chage - 2.0.0 - remove comment_html
1. comment_html is a bad name, comment_formatted would be better
2. Even so, including comment_html like this probably isn't the right thing to do.
   It also is annoying from the PHP code side of things, as we can't be sure
   a CommentFormatter is availble at the time we produce the event.

Included 'parsed comment' was originally done for other events a part of T170145.
We can add it back into our mediawiki change events if we want to, but for now it is not needed.

It is 'safe' for us to remove this field in an incompatible way, because we are still in
'development', and can manually handle this change.

Bug: T327065
Bug: T327065
Change-Id: I8f233d667d0d98372cd7dabf7c07cb59bd3ff079