Add event schema for ML classification change on current page state
mediawiki/page/prediction_classification_change schema is based
on fragment/mediawiki/state/change/page schema and represents a
classification model output on the current page state. It is a
remodelled version of mediawiki/revision/score schema, but with
the removal of the map field. This means that only one prediction
is represented per event.

Bug: T331401
Change-Id: I90f5ba25194b44513fa587903f4023ac02eee627