Encode redirect targets in page change events.
Motivation: The CirrusSearch Streaming Updater (flink) is based on the page-change event stream.
CirrusSearch does not index redirects directly but rather updates the
redirect target to keep a record of the redirect. In case of a page
delete event, there is no way to restore the redirect target from
outside the wiki issuing the event, at least not for an anonymous user
which the flink application is. Therefore, we'd like to make the
page-change events more self contained by adding information about the
redirect target of a page.

This change introduces a new MW state entity model: page_link.
As is, page_link can represent links to local or 'interwiki'
pages.  We will use this model to add more link types as we
work on T331399.

page_link is used on the mediawiki/page/change page.redirect_page_link
field to represent redirect link info.

Bug: T325315
Bug: T331399
Needed-By: I238061d64a3809fac8b66c401253fc6357dd8879
Change-Id: I5f6fe0db3db3a770c2b4d84af5d9bb2491dd530f