cirrussearch/update_pipeline/update remove required fields
In fields we were requiring the page_id and verison to exist, which
was reasonable when we had the full document available here. The
implementation has changed though, where the full document is optional
here and we will have events that contain only the weighted_tags
field. Remove the required fields to make this possible.

The version number was not changed as no events have been produced
in production yet.

Change-Id: Ifcae1d96525732f67c79419388aa08d2a0f0b1c3